requirements.txt
fastapi
uvicorn
openai
llama-index
llama-index-llms-openai
nltk
pypdf
scikit-learn
openpyxl
python-dotenv
azure-storage-blob
